Add keymaps for web frontend
