+++ src/utils/debug.ts
@@ -14,9 +14,10 @@ import {
 import { tab, settings } from '../state/state'
 
 import { format } from 'date-fns'
-import { PromptType, saveCurrentTab } from './events'
+import { PromptType } from './events/events'
 import { getBreakpoints } from './breakpoints'
 import { computed, toRaw } from 'vue'
+import { saveCurrentTab } from './events/tauri-shortcuts'
 
 export async function setBreakpoint() {
   const currentTab = tab()
