+++ src/utils/mips/mips.ts
@@ -223,17 +223,6 @@ export type InstructionLine =
   | InstructionLineComment
   | InstructionLineLabel
 
-export interface Accelerator {
-  command: boolean
-  shift: boolean
-  key: string
-}
-
-export interface Shortcut {
-  event: string
-  accelerator: Accelerator
-}
-
 export interface MipsCallbacks {
   consoleWrite(text: string, error: boolean): void
   midiPlay(note: MidiNote): void
@@ -241,8 +230,6 @@ export interface MipsCallbacks {
 
 export interface MipsBackend {
   waitReady(): Promise<void>
-
-  shortcuts(): Promise<Shortcut[]>
 
   setCallbacks(callbacks: MipsCallbacks): Promise<void>
